add absolute path support to the server.
This is handled entirely on the server side.

logger.load('/debug/some-run/variables.pkl') loads the path w.r.t.
the root log_dir for the server.

logger.load('//home/ubuntu/some-run/...') loads the system path
because of the double slash at the begining.

In the future we will add a feature on the server to limit the
read scope to sub directories of the data directory as a scoping/
saftey option.